modify

+++ dbcat/src/main/java/com/zyh/dbcat/test/IDataSupport.java
@@ -14,7 +14,7 @@ public interface IDataSupport<T> {
      * @param t
      * @return
      */
-    public int insert(T t);
+    public long insert(T t);
 
     void init(SQLiteDatabase sqLiteDatabase, Class<T> clazz);
 }
